Restore u64 to avoid panic

+++ app/packages/rust_sdk/lib/acter_flutter_sdk_ffi.dart
@@ -18519,7 +18519,7 @@ class Api {
           )>();
   late final _spaceHierarchyRoomInfoNumJoinedMembersPtr = _lookup<
       ffi.NativeFunction<
-          ffi.Uint32 Function(
+          ffi.Uint64 Function(
             ffi.Int64,
           )>>("__SpaceHierarchyRoomInfo_num_joined_members");
 
@@ -19864,7 +19864,7 @@ class Api {
           )>();
   late final _memberPowerLevelPtr = _lookup<
       ffi.NativeFunction<
-          ffi.Uint32 Function(
+          ffi.Int64 Function(
             ffi.Int64,
           )>>("__Member_power_level");
 
@@ -20109,7 +20109,7 @@ class Api {
           )>();
   late final _publicSearchResultItemNumJoinedMembersPtr = _lookup<
       ffi.NativeFunction<
-          ffi.Uint32 Function(
+          ffi.Uint64 Function(
             ffi.Int64,
           )>>("__PublicSearchResultItem_num_joined_members");
 
@@ -48656,7 +48656,7 @@ class _PublicSearchResultPrevBatchReturn extends ffi.Struct {
 class _PublicSearchResultTotalRoomCountEstimateReturn extends ffi.Struct {
   @ffi.Uint8()
   external int arg0;
-  @ffi.Uint32()
+  @ffi.Uint64()
   external int arg1;
 }
